backend/requirements.txt
aniso8601==6.0.0
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.7
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
Werkzeug==2.0.0
python-dotenv=0.17.1
